+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/SnapshotDslTest.java
@@ -101,6 +101,11 @@ public class SnapshotDslTest {
             return this;
         }
 
+        @Override
+        public SnapshotTestResult disabledBecause(String message) {
+            return null;
+        }
+
     }
 
 }